Rename directory 'structure' to 'head'
