+++ creator/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './structure/app';
+import App from './main/app';
 import './index.css';
 
 ReactDOM.render(
